Remove require command from WC_EBANX_Hooks and start use a namespace
